requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
cffi==1.15.1
Django==3.2.15
django-debug-toolbar==3.6.0
djangorestframework==3.13.1
psycopg2==2.9.3
pycparser==2.21
pytz==2022.2.1
sqlparse==0.4.2
typing-extensions==4.3.0
